requirements.txt
openai==1.97.1
pytesseract
pdfplumber
pillow
PyPDF2
faiss-cpu
pdf2image